layout/base/tests/test_bug761572.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=761572
-->
<head>
  <title>Test for Bug 761572</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=761572">Mozilla Bug 761572</a>
<div id="content">
  <div id="d" style="background:lime; width:50px; height:50px" onmouseup="doUp()" onclick="doClick()"></div>
</div>
<pre id="test">
<script type="application/javascript">
SimpleTest.waitForExplicitFinish();

var d = document.getElementById("d");

function doUp() {
  d.style.display = "none";
}
function doClick() {
  ok(true, "Check click received");
  SimpleTest.finish();
}
  
function doTest() {
  // synthesizes a mousedown/mouseup pair
  synthesizeMouse(d, 10, 10, {});
}

SimpleTest.waitForFocus(doTest);
</script>
</pre>
</body>
</html>
